+++ src/components/globals/Track.tsx
@@ -1,5 +1,7 @@
 import styled from "styled-components";
 
+import { ReactComponent as PlaySVG } from "@assets/icons/ic_30_play_point.svg";
+
 import { T5Medium, T6Medium } from "@components/Typography";
 
 import colors from "@constants/colors";
@@ -8,11 +10,16 @@ interface TrackProps {
   // TODO: Interface 작업 예정
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   item: any;
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  onClick?: (item: any) => void;
 }
 
-const Track = ({ item }: TrackProps) => {
+const Track = ({ item, onClick }: TrackProps) => {
   return (
-    <Container>
+    <Container onClick={() => onClick && onClick(item)}>
+      <PlayIcon />
+
       <Thumbnail src={`https://i.ytimg.com/vi/${item.songId}/hqdefault.jpg`} />
 
       <TrackInfo>
@@ -23,12 +30,13 @@ const Track = ({ item }: TrackProps) => {
   );
 };
 
-const Container = styled.div`
-  display: flex;
-  align-items: center;
+const PlayIcon = styled(PlaySVG)`
+  display: none;
+
+  z-index: 1;
 `;
 
-export const Thumbnail = styled.img`
+const Thumbnail = styled.img`
   margin-right: 8px;
 
   width: 78px;
@@ -36,6 +44,27 @@ export const Thumbnail = styled.img`
 
   object-fit: cover;
   border-radius: 4px;
+`;
+
+const Container = styled.div`
+  position: relative;
+
+  display: flex;
+  align-items: center;
+
+  cursor: pointer;
+
+  &:hover ${Thumbnail} {
+    filter: brightness(0.4);
+  }
+
+  &:hover ${PlayIcon} {
+    display: block;
+
+    position: absolute;
+    left: 18px;
+    top: 7px;
+  }
 `;
 
 const TrackInfo = styled.div``;
