+++ src/components/index/ChartItem.tsx
@@ -1,9 +1,7 @@
 import styled from "styled-components";
 
-import { ReactComponent as PlaySVG } from "@assets/icons/ic_30_play_point.svg";
-
 import Rank from "@components/globals/Rank";
-import Track, { Thumbnail } from "@components/globals/Track";
+import Track from "@components/globals/Track";
 
 interface ChartItemProps {
   rank: number;
@@ -17,15 +15,10 @@ const ChartItem = ({ rank, item }: ChartItemProps) => {
   return (
     <Container>
       <Rank now={rank} last={item.hourly.last} />
-      <Track item={item} />
-      <PlayIcon />
+      <Track item={item} onClick={console.log} />
     </Container>
   );
 };
-
-const PlayIcon = styled(PlaySVG)`
-  display: none;
-`;
 
 const Container = styled.div`
   position: relative;
@@ -35,20 +28,6 @@ const Container = styled.div`
 
   display: flex;
   align-items: center;
-
-  cursor: pointer;
-
-  &:hover ${Thumbnail} {
-    filter: brightness(0.4);
-  }
-
-  &:hover ${PlayIcon} {
-    display: block;
-
-    position: absolute;
-    left: 63px;
-    top: 17px;
-  }
 `;
 
 export default ChartItem;
